+++ node/src/types/kernel/api/notebook/setNotebookConf/index.d.ts
@@ -17,13 +17,7 @@
 
 
 /* directories */
-export * as closeNotebook from "./closeNotebook";
-export * as createNotebook from "./createNotebook";
-export * as getNotebookConf from "./getNotebookConf";
-export * as lsNotebooks from "./lsNotebooks";
-export * as openNotebook from "./openNotebook";
-export * as removeNotebook from "./removeNotebook";
-export * as renameNotebook from "./renameNotebook";
-export * as setNotebookConf from "./setNotebookConf";
 
 /* flles */
+export * from "./payload";
+export * from "./response";