+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - 添加 API `/api/notebook/renameNotebook` | Add API `/api/notebook/renameNotebook`
 - 添加 API `/api/notebook/createNotebook` | Add API `/api/notebook/createNotebook`
 - 添加 API `/api/notebook/removeNotebook` | Add API `/api/notebook/removeNotebook`
+- 添加 API `/api/notebook/setNotebookConf` | Add API `/api/notebook/setNotebookConf`
 
 ## v0.0.4 / 2023-06-29